requirements.txt
alembic==1.17.2
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asn1crypto==1.5.1
babel==2.17.0
bcrypt==5.0.0
bleach==6.3.0
blinker==1.9.0
cbor2==5.7.1
cffi==2.0.0
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
dnspython==2.8.0
email-validator==2.3.0
Flask==3.1.2
flask-babel==4.0.0
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Principal==0.4.0
Flask-Security-Too==5.7.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.4
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
libpass==1.9.3
Mako==1.3.10
MarkupSafe==3.0.3
packaging==25.0
phonenumberslite==9.0.20
pluggy==1.6.0
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.3.0
pytest==9.0.2
pytz==2025.2
qrcode==8.2
SQLAlchemy==2.0.44
typing_extensions==4.15.0
webauthn==2.7.0
webencodings==0.5.1
Werkzeug==3.1.3
WTForms==3.2.1
